extract histogram implementation
